hoc for Buttons added | created questions page
