+++ client/src/hoc/withIconButton/style.scss
@@ -1,0 +1,11 @@
+.questionButton-container {
+  height: 66px;
+  width: 66px;
+  background-color: white;
+  border: 0 solid;
+  border-radius: 50%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  cursor: pointer;
+}
