+++ client/src/pages/Questions/style.scss
@@ -1,0 +1,152 @@
+.home {
+  &__container {
+    display: flex;
+    justify-content: center;
+    flex-wrap: nowrap;
+    margin-top: 25px;
+    padding: 0 30px;
+  }
+
+  &__card {
+    position: relative;
+  }
+
+  &__img__page {
+    height: 550px;
+    width: 318px;
+    object-fit: cover;
+    object-position: center;
+    border-radius: 19px;
+    overflow: hidden;
+  }
+
+  &__card_circle_container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    position: absolute;
+    bottom: 16px;
+    left: 0;
+    color: white;
+    width: 100%;
+  }
+
+  &__card__message {
+    h2 {
+      font-size: 18px;
+      text-align: center;
+      margin-bottom: 4px;
+      font-weight: lighter;
+    }
+    h1 {
+      text-transform: uppercase;
+      font-weight: bold;
+      font-size: 24px;
+      margin-bottom: 8px;
+      text-align: center;
+    }
+  }
+
+  &__circle__container {
+    display: flex;
+    justify-content: space-evenly;
+    align-items: center;
+    width: 100%;
+  }
+
+  &__circle {
+    height: 66px;
+    width: 66px;
+    background-color: white;
+    border-radius: 50%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    cursor: pointer;
+  }
+
+  &__message {
+    display: none;
+    flex-grow: 1;
+  }
+
+  @media screen and (min-width: 768px) {
+    &__img__page {
+      width: 100%;
+      border-radius: 19px 0 0 19px;
+    }
+
+    &__card__message {
+      display: none;
+    }
+
+    &__message {
+      display: flex;
+      background-color: rgba(251, 251, 251, 1);
+      border-radius: 0 19px 19px 0;
+      justify-content: center;
+      align-items: center;
+      font-family: 'Helvetica Neue', sans-serif;
+      color: rgba(36, 36, 36, 1);
+
+      div {
+        display: flex;
+        flex-direction: column;
+
+        div {
+          margin-top: 11px;
+        }
+
+        h2 {
+          font-size: 25px;
+          font-weight: lighter;
+        }
+
+        p {
+          font-size: 18px;
+          color: rgba(78, 78, 78, 1);
+          text-align: center;
+          margin-top: 5px;
+        }
+
+        h1{
+          text-transform: uppercase;
+          font-size: 31px;
+          font-weight: bold;
+          margin-top: 5px;
+        }
+
+      }
+
+    }
+
+    @media screen and (min-width: 992px) {
+      &__message div {
+
+        div {
+          margin-top: 18px;
+        }
+
+        h2 {
+          font-size: 30px;
+        }
+
+        p {
+          font-size: 23px;
+        }
+
+        h1 {
+          font-size: 36px;
+          margin-top: 10px;
+        }
+
+      }
+    }
+
+  }
+
+}
+
+
+
